requirements.txt
torch
opencv-python
allure-pytest
pytest
numpy
scikit-learn
matplotlib
pillow
pytest-html 